+++ backend/DataBaseTrello/API/DTO/Mappers/ToDomainMapper.cs
@@ -1,9 +1,9 @@
 ﻿using API.DTO.Domain;
 using API.DTO.Requests;
 
-namespace API.DTO.Mappers.ToDomainModel
+namespace API.DTO.Mappers
 {
-    public static class ToDomainModelMapper
+    public static class ToDomainMapper
     {
         public static UpdateUserModel ToUpdateUserModel(UpdateUserRequest request)
         {
