+++ backend/DataBaseTrello/API/Services/UserService.cs
@@ -3,8 +3,7 @@ using System.Net;
 
 using API.Constants;
 using API.DTO.Domain;
-using API.DTO.Mappers.ToResponseModel;
-using API.DTO.Responses;
+using API.DTO.Mappers;
 using API.Exceptions;
 using API.Exceptions.Context;
 using API.Extensions;
@@ -73,8 +72,8 @@ namespace DataBaseInfo.Services
             using var context = _contextFactory.CreateDbContext();
          
             User? user = await context.Users.FirstOrDefaultAsync(u => u.UserEmail == UserEmail);
-                    if (user == null)
-                    throw new AppException(_errCreator.Unauthorized($"Учётной записи с Email: {UserEmail} не существует"));
+            if (user == null)
+                throw new AppException(_errCreator.Unauthorized($"Учётной записи с Email: {UserEmail} не существует"));
         
             var activeToken = await context.RefreshTokens
             .Include(rt => rt.User)
@@ -83,46 +82,38 @@ namespace DataBaseInfo.Services
                 && !rt.IsRevoked
                 && rt.ExpiresAt > DateTime.UtcNow);
 
-                    if (activeToken != null)
-                    throw new AppException(_errCreator.Conflict($"Пользователь id: {user.Id}, email: {UserEmail}. Уже был авторизован"));
+             if (activeToken != null)
+                throw new AppException(_errCreator.Conflict($"Пользователь id: {user.Id}, email: {UserEmail}. Уже был авторизован"));
           
             var result = new PasswordHasher<User?>().VerifyHashedPassword(user, user.UserPassword, Password);
-                    if (result == PasswordVerificationResult.Success)
-                    {
-                        var refreshToken = Guid.NewGuid().ToString();
-             
-                        await _JWTService.CreateRefreshTokenAsync(user, refreshToken);
-                        var accessToken = _JWTService.GenerateAccessToken(user);
-                        return (accessToken, refreshToken);
 
-                    }
-                    else
-                    {
-                    throw new AppException(_errCreator.Unauthorized($"Неверно введён пароль к аккаунту id: {user.Id}, email:{UserEmail}"));
-                }
+            if (result != PasswordVerificationResult.Success)
+                throw new AppException(_errCreator.Unauthorized($"Неверно введён пароль к аккаунту id: {user.Id}, email:{UserEmail}"));
             
+            var refreshToken = Guid.NewGuid().ToString();
+             
+            await _JWTService.CreateRefreshTokenAsync(user, refreshToken);
+            var accessToken = _JWTService.GenerateAccessToken(user);
+            return (accessToken, refreshToken);
         }
 
         public async Task<string> UpdateUserAvatarAsync(Result? result, int userId)
         {
-            if (result.HttpStatusCode >= 200 && result.HttpStatusCode < 300)
-            {
-                using var context = await _contextFactory.CreateDbContextAsync();
-
-                var user = await context.Users.FirstOrDefaultAsync(u => u.Id == userId);
-
-                if (user == null)
-                    throw new AppException(_errCreator.NotFound($"Произошла ошибка в момент смены аватара пользователя, Пользователь id: {userId}, не найден"));
-
-                user.Avatar = result.url;
-                await context.SaveChangesWithContextAsync($"Произошла ошибка в момент смены аватара пользователя, не удалось сохранить url: {result.url}");
-
-                return result.url;
-            }
-            else
-            {
+            if (result.HttpStatusCode >= 400 && result.HttpStatusCode <= 500)
                 throw new AppException(_errCreator.InternalServerError($"Ошибка при загрузке изображения в ImageKit. Код: {result.HttpStatusCode}"));
-            }
+
+            using var context = await _contextFactory.CreateDbContextAsync();
+
+            var user = await context.Users.FirstOrDefaultAsync(u => u.Id == userId);
+
+            if (user == null)
+                throw new AppException(_errCreator.NotFound($"Произошла ошибка в момент смены аватара пользователя, Пользователь id: {userId}, не найден"));
+
+            user.Avatar = result.url;
+            await context.SaveChangesWithContextAsync($"Произошла ошибка в момент смены аватара пользователя, не удалось сохранить url: {result.url}");
+
+            return result.url;
+          
         }
         public async Task<UpdateUserModel> UpdateUserAsync(UpdateUserModel model, int userId)
         {
@@ -131,13 +122,8 @@ namespace DataBaseInfo.Services
 
             if (user == null)
                 throw new AppException(_errCreator.NotFound($"Ошибка при получении данных, user {userId} не найден"));
-            
-            if(!string.IsNullOrEmpty(model.FirstUserName))
-                user.FirstName = model.FirstUserName;
-            if (!string.IsNullOrEmpty(model.LastUserName))
-                user.SecondName = model.LastUserName;
-            if (model.Sex!=null)
-                user.Sex = (Sex)model.Sex;
+
+            ToEntityMapper.ApplyToUser(model, user);
 
             await context.SaveChangesWithContextAsync($"Ошибка при обновлении данных о пользователе: {userId}");
